extract DrtModeOptimizerQSimModule
